+++ v6_version.py
@@ -1,3 +1,3 @@
 """ASL V6 release version."""
 
-__version__ = "6.1.1"
+__version__ = "6.1.2"